website/familynetapp/serializers.py
from rest_framework import serializers

from familynetapp.models import Person, Relation


class PersonSerializer(serializers.ModelSerializer):
    class Meta:
        model = Person
        fields = ('id', 'first_name', 'last_name')


class RelationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Relation
        fields = ('id', 'source', 'target', 'relationship')
